move process status enum to repo dir
