+++ rest/src/main/java/my/raptive/url/rest/repository/ProcessStatus.java
@@ -1,4 +1,4 @@
-package my.raptive.url.rest;
+package my.raptive.url.rest.repository;
 
 public enum ProcessStatus {
     RECEIVED, SUCCESS, FAILED, IN_PROGRESS
